Inspect backup contents - Implements backup inspection controller and workers - Adds new --inspect-increments command line option - Adds new --inspect-content command line option - Adds new tests - Updates documentation
Resolves #110
{minor}

+++ file-barj-core/src/main/java/com/github/nagyesta/filebarj/core/restore/pipeline/RestoreController.java
@@ -50,7 +50,7 @@ public class RestoreController {
         final ManifestManager manifestManager = new ManifestManagerImpl();
         //TODO: allow restoring earlier versions
         log.info("Loading backup manifests for restore from: {}", backupDirectory);
-        final var manifests = manifestManager.load(backupDirectory, fileNamePrefix, kek, Integer.MAX_VALUE);
+        final var manifests = manifestManager.load(backupDirectory, fileNamePrefix, kek, Long.MAX_VALUE);
         log.info("Merging {} manifests", manifests.size());
         manifest = manifestManager.mergeForRestore(manifests);
         final var filesOfLastManifest = manifest.getFilesOfLastManifest();
